Mon, Oct 16, 2023, 9:52 AM -05:00

+++ llama.c
@@ -7,7 +7,7 @@ int main()
     year = 0;
     n1 = 2;
     n2 = 3;
-    int change = 0;
+    double change = 0.000;
     int n3;
     do {
         printf("\nPlease enter the starting population size of llamas, ensuring your input is a number greater than 9 and less than 10,000: ");
@@ -47,7 +47,7 @@ int main()
         change = born - dead;
         n3 = n2 - n1;
         printf("%d\n", n3);
-        printf("%d", change);
+        printf("%f", change);
         }
     }
     while (n2 < n1);
